+++ apps/landing-admin-web/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @client/web
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @southneuhof/is-vue-framework@1.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- @client/section-schema@0.0.1
+
 ## 0.0.2
 
 ### Patch Changes
@@ -10,7 +23,7 @@
   - @southneuhof/utilities@0.1.1
   - @southneuhof/is-vue-framework@1.0.1
   - @client/data-model@0.0.0
-  - @southneuhof/landing-section-schema@0.0.2
+  - @client/section-schema@0.0.2
 
 ## 0.0.1
 
@@ -19,4 +32,4 @@
 - Updated dependencies [92ab493]
   - @southneuhof/utilities@0.1.0
   - @southneuhof/is-vue-framework@1.0.0
-  - @southneuhof/landing-section-schema@0.0.1
+  - @client/section-schema@0.0.1